refactor: delete passenger service interface, delete ovveride annotations from custom exceptions classes methods and update soft-delete method
