+++ src/main/java/com/software/modsen/passengermicroservice/exceptions/PassengerNotFoundException.java
@@ -5,7 +5,6 @@ public class PassengerNotFoundException extends RuntimeException {
         super(message);
     }
 
-    @Override
     public String getMessage() {
         return super.getMessage();
     }
