+++ src/main/java/com/software/modsen/passengermicroservice/exceptions/PassengerWasDeletedException.java
@@ -5,7 +5,6 @@ public class PassengerWasDeletedException extends RuntimeException {
         super(message);
     }
 
-    @Override
     public String getMessage() {
         return super.getMessage();
     }
